# src/main/java/com/cmsDemo/app/api/SectionsApi.java
/**
 * NOTE: This class is auto generated by the swagger code generator program (2.4.21).
 * https://github.com/swagger-api/swagger-codegen
 * Do not edit the class manually.
 */
package com.cmsDemo.app.api;

import com.cmsDemo.app.model.Section;
import com.cmsDemo.app.model.Teacher;
import io.swagger.annotations.*;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestHeader;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestMethod;
import org.springframework.web.bind.annotation.RequestParam;
import org.springframework.web.bind.annotation.RequestPart;
import org.springframework.web.multipart.MultipartFile;

import javax.validation.Valid;
import javax.validation.constraints.*;
import java.util.List;
@javax.annotation.Generated(value = "io.swagger.codegen.languages.SpringCodegen", date = "2021-08-20T11:36:07.743+05:30")

@Validated
@Api(value = "sections", description = "the sections API")
@RequestMapping(value = "")
public interface SectionsApi {

    @ApiOperation(value = "", nickname = "createSection", notes = "Create new section", tags={  })
    @ApiResponses(value = { 
        @ApiResponse(code = 200, message = "Successful operation"),
        @ApiResponse(code = 409, message = "Section with id already exists") })
    @RequestMapping(value = "/sections",
        consumes = { "application/json" },
        method = RequestMethod.POST)
    ResponseEntity<Void> createSection(@ApiParam(value = "section object" ,required=true )  @Valid @RequestBody Section body);


    @ApiOperation(value = "", nickname = "deleteSection", notes = "Delete a section using ID", tags={  })
    @ApiResponses(value = { 
        @ApiResponse(code = 200, message = "Successful operation"),
        @ApiResponse(code = 404, message = "Section not found") })
    @RequestMapping(value = "/sections/{sectionId}",
        method = RequestMethod.DELETE)
    ResponseEntity<Void> deleteSection(@ApiParam(value = "ID of section to return",required=true) @PathVariable("sectionId") String sectionId);


    @ApiOperation(value = "", nickname = "getAllsections", notes = "Returns a list of sections", response = Section.class, responseContainer = "List", tags={  })
    @ApiResponses(value = { 
        @ApiResponse(code = 200, message = "Successful operation", response = Section.class, responseContainer = "List"),
        @ApiResponse(code = 204, message = "No sections in DB") })
    @RequestMapping(value = "/sections",
        produces = { "application/json" }, 
        method = RequestMethod.GET)
    ResponseEntity<List<Section>> getAllSections();


    @ApiOperation(value = "", nickname = "getSectionById", notes = "Get a section using ID", response = Section.class, tags={  })
    @ApiResponses(value = { 
        @ApiResponse(code = 200, message = "Successful operation", response = Section.class),
        @ApiResponse(code = 400, message = "Invalid ID supplied"),
        @ApiResponse(code = 404, message = "Section not found") })
    @RequestMapping(value = "/sections/{sectionId}",
        produces = { "application/json" }, 
        method = RequestMethod.GET)
    ResponseEntity<Section> getSectionById(@ApiParam(value = "ID of section to return",required=true) @PathVariable("sectionid") String sectionid);


    @ApiOperation(value = "", nickname = "updateSection", notes = "Update section", tags={  })
    @ApiResponses(value = { 
        @ApiResponse(code = 200, message = "Successful operation"),
        @ApiResponse(code = 404, message = "Section with id does not exists") })
    @RequestMapping(value = "/sections",
        consumes = { "application/json" },
        method = RequestMethod.PUT)
    ResponseEntity<Void> updateSection(@ApiParam(value = "section object" ,required=true )  @Valid @RequestBody Section body);

}